# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KCardSwap API",
    "description": "Backend API for KCardSwap - Card Exchange Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ratings": {
      "post": {
        "tags": [
          "Ratings"
        ],
        "summary": "Submit rating",
        "description": "Submit a rating for another user after a trade",
        "operationId": "submit_rating_api_v1_ratings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Rating submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (validation failed)"
          },
          "401": {
            "description": "Unauthorized (not logged in)"
          },
          "422": {
            "description": "Unprocessable entity (cannot rate user)"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ratings/user/{user_id}": {
      "get": {
        "tags": [
          "Ratings"
        ],
        "summary": "Get user ratings",
        "description": "Get ratings received by a specific user",
        "operationId": "get_user_ratings_api_v1_ratings_user__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of ratings",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of ratings"
          }
        ],
        "responses": {
          "200": {
            "description": "Ratings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (not logged in)"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ratings/user/{user_id}/average": {
      "get": {
        "tags": [
          "Ratings"
        ],
        "summary": "Get average rating",
        "description": "Get average rating score for a specific user",
        "operationId": "get_average_rating_api_v1_ratings_user__user_id__average_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Average rating retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AverageRatingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (not logged in)"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Get my reports",
        "description": "Get reports submitted by the current user",
        "operationId": "get_my_reports_api_v1_reports_get",
        "responses": {
          "200": {
            "description": "Reports retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (not logged in)"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Reports"
        ],
        "summary": "Submit report",
        "description": "Submit a report for inappropriate behavior or content",
        "operationId": "submit_report_api_v1_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Report submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (validation failed)"
          },
          "401": {
            "description": "Unauthorized (not logged in)"
          },
          "422": {
            "description": "Unprocessable entity (cannot report user)"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AverageRatingResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "User ID"
          },
          "average_score": {
            "type": "number",
            "title": "Average Score",
            "description": "Average rating score",
            "examples": [
              4.8
            ]
          },
          "total_ratings": {
            "type": "integer",
            "title": "Total Ratings",
            "description": "Total number of ratings received"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "average_score",
          "total_ratings"
        ],
        "title": "AverageRatingResponse",
        "description": "Response schema for average rating",
        "example": {
          "average_score": 4.8,
          "total_ratings": 15,
          "user_id": "123e4567-e89b-12d3-a456-426614174000"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RatingListResponse": {
        "properties": {
          "ratings": {
            "items": {
              "$ref": "#/components/schemas/RatingResponse"
            },
            "type": "array",
            "title": "Ratings",
            "description": "List of ratings"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of ratings"
          }
        },
        "type": "object",
        "required": [
          "ratings",
          "total"
        ],
        "title": "RatingListResponse",
        "description": "Response schema for list of ratings",
        "example": {
          "ratings": [
            {
              "comment": "Great trader!",
              "created_at": "2024-01-01T00:00:00Z",
              "id": "456e7890-e12b-34d5-a678-901234567000",
              "rated_user_id": "123e4567-e89b-12d3-a456-426614174000",
              "rater_id": "789e0123-e45b-67d8-a901-234567890000",
              "score": 5,
              "trade_id": "987e6543-e21b-12d3-a456-426614174000"
            }
          ],
          "total": 1
        }
      },
      "RatingRequest": {
        "properties": {
          "rated_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Rated User Id",
            "description": "ID of user being rated",
            "examples": [
              "123e4567-e89b-12d3-a456-426614174000"
            ]
          },
          "trade_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Id",
            "description": "Associated trade ID (optional)",
            "examples": [
              "987e6543-e21b-12d3-a456-426614174000"
            ]
          },
          "score": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Score",
            "description": "Rating score (1-5)",
            "examples": [
              5
            ]
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Optional comment",
            "examples": [
              "Great trader! Fast and friendly communication."
            ]
          }
        },
        "type": "object",
        "required": [
          "rated_user_id",
          "score"
        ],
        "title": "RatingRequest",
        "description": "Request schema for submitting a rating",
        "example": {
          "comment": "Great trader! Fast and friendly communication.",
          "rated_user_id": "123e4567-e89b-12d3-a456-426614174000",
          "score": 5,
          "trade_id": "987e6543-e21b-12d3-a456-426614174000"
        }
      },
      "RatingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Rating ID"
          },
          "rater_id": {
            "type": "string",
            "format": "uuid",
            "title": "Rater Id",
            "description": "User who gave the rating"
          },
          "rated_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Rated User Id",
            "description": "User who received the rating"
          },
          "trade_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Id",
            "description": "Associated trade ID"
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "description": "Rating score (1-5)"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Rating comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rater_id",
          "rated_user_id",
          "score",
          "created_at"
        ],
        "title": "RatingResponse",
        "description": "Response schema for a rating",
        "example": {
          "comment": "Great trader! Fast and friendly communication.",
          "created_at": "2024-01-01T00:00:00Z",
          "id": "456e7890-e12b-34d5-a678-901234567000",
          "rated_user_id": "123e4567-e89b-12d3-a456-426614174000",
          "rater_id": "789e0123-e45b-67d8-a901-234567890000",
          "score": 5,
          "trade_id": "987e6543-e21b-12d3-a456-426614174000"
        }
      },
...
      "ReportListResponse": {
        "properties": {
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportResponse"
            },
            "type": "array",
            "title": "Reports",
            "description": "List of reports"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of reports"
          }
        },
        "type": "object",
        "required": [
          "reports",
          "total"
        ],
        "title": "ReportListResponse",
        "description": "Response schema for list of reports",
        "example": {
          "reports": [
            {
              "created_at": "2024-01-01T00:00:00Z",
              "detail": "User sent inappropriate messages.",
              "id": "456e7890-e12b-34d5-a678-901234567000",
              "reason": "Inappropriate behavior",
              "reported_user_id": "123e4567-e89b-12d3-a456-426614174000",
              "reporter_id": "789e0123-e45b-67d8-a901-234567890000",
              "status": "pending"
            }
          ],
          "total": 1
        }
      },
      "ReportRequest": {
        "properties": {
          "reported_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reported User Id",
            "description": "ID of user being reported",
            "examples": [
              "123e4567-e89b-12d3-a456-426614174000"
            ]
          },
          "reason": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Reason",
            "description": "Reason for report",
            "examples": [
              "Inappropriate behavior",
              "Spam",
              "Scam",
              "Harassment"
            ]
          },
          "detail": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Detailed description of the issue",
            "examples": [
              "User sent inappropriate messages and refused to complete trade."
            ]
          }
        },
        "type": "object",
        "required": [
          "reported_user_id",
          "reason"
        ],
        "title": "ReportRequest",
        "description": "Request schema for submitting a report",
        "example": {
          "detail": "User sent inappropriate messages and refused to complete trade.",
          "reason": "Inappropriate behavior",
          "reported_user_id": "123e4567-e89b-12d3-a456-426614174000"
        }
      },
      "ReportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Report ID"
          },
          "reporter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reporter Id",
            "description": "User who filed the report"
          },
          "reported_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reported User Id",
            "description": "User who was reported"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Reason for report"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Detailed description"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Report status",
            "examples": [
              "pending",
              "reviewed",
              "resolved"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reporter_id",
          "reported_user_id",
          "reason",
          "status",
          "created_at"
        ],
        "title": "ReportResponse",
        "description": "Response schema for a report",
        "example": {
          "created_at": "2024-01-01T00:00:00Z",
          "detail": "User sent inappropriate messages.",
          "id": "456e7890-e12b-34d5-a678-901234567000",
          "reason": "Inappropriate behavior",
          "reported_user_id": "123e4567-e89b-12d3-a456-426614174000",
          "reporter_id": "789e0123-e45b-67d8-a901-234567890000",
          "status": "pending"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}